requirement.txt
surya-ocr
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 
gspread